#11 Implement: Sell Products List 화면 수정
1. Fix: AddEditProductActivity ADD_EDIT Request code split -> ADD, EDIT
2. Implement: 수정된 디자인에 맞춰 뷰 다시 구현

+++ app/src/main/java/kr/co/mash_up/a9tique/ui/sellproducts/SellProductsPresenter.java
@@ -113,6 +113,7 @@ public class SellProductsPresenter implements SellProductsContract.Presenter {
                         if (showLoadingUI) {
                             mView.setLodingIndicator(false);
                         }
+                        mView.showNoProducts();
                         mView.showLoadingProductsError();
                     }
                 });
@@ -197,7 +198,11 @@ public class SellProductsPresenter implements SellProductsContract.Presenter {
     }
 
     @Override
-    public void removeProductAll(List<Product> products) {
+    public void removeProductSelected(List<Product> products) {
+        if (products.size() == 0) {
+            mView.showRemoveProductSelectedErrorMessage();
+            return;
+        }
         mView.setProgressbar(true);
 
         BackendHelper.getInstance().deleteSellProducts(products, new ResultCallback() {
